+++ apps/activity-log/src/payload/globals/GlobalWithDrafts.ts
@@ -1,0 +1,33 @@
+import type { GlobalConfig } from "payload";
+
+export const GlobalWithDrafts: GlobalConfig = {
+  slug: "global-with-drafts",
+  fields: [
+    {
+      name: "text",
+      type: "text",
+      required: true,
+    },
+    {
+      name: "array",
+      type: "array",
+      fields: [
+        {
+          name: "key",
+          type: "text",
+        },
+        {
+          name: "value",
+          type: "text",
+        },
+      ],
+    },
+  ],
+  versions: {
+    drafts: {
+      autosave: {
+        interval: 100,
+      },
+    },
+  },
+};