+++ packages/soft-delete/src/endpoints/index.ts
@@ -23,8 +23,10 @@ export const endpoints: Endpoint[] = [
           collection: data?.["collection"],
           id: data["id"],
           data: {
+            // @ts-expect-error payload trash feature has broken this type
             deletedAt: new Date(),
           },
+          req,
         });
       } else if (data?.["ids"]) {
         response = await req.payload.update({
@@ -35,8 +37,10 @@ export const endpoints: Endpoint[] = [
             },
           },
           data: {
+            // @ts-expect-error payload trash feature has broken this type
             deletedAt: new Date(),
           },
+          req,
         });
       }
 
@@ -67,6 +71,7 @@ export const endpoints: Endpoint[] = [
         response = await req.payload.delete({
           collection: data?.["collection"],
           id: data["id"],
+          req,
         });
       } else if (data?.["ids"]) {
         response = await req.payload.delete({
@@ -76,6 +81,7 @@ export const endpoints: Endpoint[] = [
               in: data["ids"],
             },
           },
+          req,
         });
       }
 
@@ -109,6 +115,7 @@ export const endpoints: Endpoint[] = [
           data: {
             deletedAt: null,
           },
+          req,
         });
       } else if (data?.["ids"]) {
         response = await req.payload.update({
@@ -121,6 +128,7 @@ export const endpoints: Endpoint[] = [
           data: {
             deletedAt: null,
           },
+          req,
         });
       }