+++ packages/broken-link-checker/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @payload-bites/broken-link-checker
 
+## 1.0.0-beta.2
+
+### Patch Changes
+
+- fix: formatting ([bc30832](https://github.com/rilrom/payload-bites/commit/bc30832))
+- chore: upgrade dependencies ([f6dad0e](https://github.com/rilrom/payload-bites/commit/f6dad0e))
+
 ## 1.0.0-beta.1
 
 ### Major Changes